+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-## [Unreleased]
+## Changelog
 
-## [0.1.0] - 2025-03-16
+## [0.2.0] - 2025-03-18
+### Added
+- Added square cool name generator
+
+## [0.1.1] - 2025-03-16
 
 - Initial release
+- Add alphanumeric cool name generator